Introduce interfaces for dispatching git commands
A handler can implement any combination of the various interfaces. If a
command is invoked on the server and the handler is compatible with that
command, the command will be dispatched to the handler for further
processing. Right now, only the `UploadPack` and `ReceivePack`
interfaces exist and only `UploadPack` has a basic implementation.

The handler can be customized as part of the config.